requirements.txt
alembic==1.4.3
altgraph==0.17
APScheduler==3.6.3
Babel==2.8.0
blinker==1.4
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.3
click==7.1.2
cssselect2==0.3.0
defusedxml==0.6.0
dnspython==2.0.0
docxtpl==0.9.2
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WeasyPrint==0.6
Flask-WTF==0.14.3
future==0.18.2
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.2
pdfkit==0.6.1
pefile==2019.4.18
Pillow==7.2.0
portalocker==2.0.0
pycparser==2.20
PyInstaller==3.6
PyJWT==1.7.1
PyPDF2==1.26.0
Pyphen==0.9.5
PySimpleGUI==4.19.0
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
six==1.14.0
SQLAlchemy==1.3.19
tinycss2==1.0.2
tzlocal==2.1
visitor==0.1.3
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
wxPython==4.1.0
youtube-dl==2020.5.3
